Put layout js into a module
